+++ tests/Service/Article/ArticleServiceTest.php
@@ -3,6 +3,7 @@
 namespace Speicher210\Fastbill\Test\Api\Service\Article;
 
 use Speicher210\Fastbill\Api\Model\Article;
+use Speicher210\Fastbill\Api\Model\Customer;
 use Speicher210\Fastbill\Api\Model\Feature;
 use Speicher210\Fastbill\Api\Model\Translation;
 use Speicher210\Fastbill\Api\Model\TranslationText;
@@ -17,19 +18,75 @@ use Speicher210\Fastbill\Test\Api\Service\AbstractServiceTest;
 class ArticleServiceTest extends AbstractServiceTest
 {
 
-    public function testGetArticle()
+    public function testGetArticles()
     {
         /** @var ArticleService $articleService */
         $articleService = $this->getServiceToTest();
-        $apiResponse = $articleService->getArticles(666);
+        $apiResponse = $articleService->getArticles(1);
 
         $this->assertInstanceOf(GetApiResponse::class, $apiResponse);
         /** @var GetResponse $response */
         $response = $apiResponse->getResponse();
-        $this->assertFalse($response->hasErrors());
 
+        $this->assertEquals(array($this->getExpectedArticle()), $response->getArticles());
+    }
+
+    public function testGetArticle()
+    {
+        /** @var ArticleService $articleService */
+        $articleService = $this->getServiceToTest();
+
+        $actualArticle = $articleService->getArticle(1);
+        $expectedArticle = $this->getExpectedArticle();
+
+        $this->assertEquals($expectedArticle, $actualArticle);
+    }
+
+    public function testGetCheckoutURLThrowsExceptionIfArticleIsNotFound()
+    {
+        /** @var ArticleService $articleService */
+        $articleService = $this->getServiceToTest();
+
+        $this->setExpectedException('\OutOfBoundsException', 'Article not found.');
+
+        $articleService->getCheckoutURL('NON-EXISTING');
+    }
+
+    public function testGetCheckoutURLReturnsTheURLIfNoCustomerPassed()
+    {
+        /** @var ArticleService $articleService */
+        $articleService = $this->getServiceToTest();
+
+        $this->assertSame(
+            'https://automatic.fastbill.com/purchase/aa9122707e4baf2090e23babe7473a79/1',
+            $articleService->getCheckoutURL(1)
+        );
+    }
+
+    public function testGetCheckoutURLReturnsTheURLForACustomer()
+    {
+        /** @var ArticleService $articleService */
+        $articleService = $this->getServiceToTest();
+
+        $customer = new Customer();
+        $customer->setHash('customer-hash');
+
+
+        $this->assertSame(
+            'https://automatic.fastbill.com/checkout/0/account-has/customer-hash/1',
+            $articleService->getCheckoutURL('1', $customer)
+        );
+    }
+
+    /**
+     * Get an article for expected.
+     *
+     * @return Article
+     */
+    private function getExpectedArticle()
+    {
         $expectedArticle = new Article();
-        $expectedArticle->setArticleNumber(666);
+        $expectedArticle->setArticleNumber(1);
         $expectedArticle->setTitle('Article test');
         $expectedArticle->setDescription('Article test description');
         $expectedArticle->setTags(array('tag1', 'tag2'));
@@ -53,11 +110,14 @@ class ArticleServiceTest extends AbstractServiceTest
         $expectedArticle->setSubscriptionCancellationPeriod('1 day');
         $expectedArticle->setReturnUrlSuccess('https://test.com/success');
         $expectedArticle->setReturnUrlCancel('https://test.com/cancel');
-        $expectedArticle->setCheckoutUrl('https://automatic.fastbill.com/purchase/aa9122707e4baf2090e23babe7473a79/666');
+        $expectedArticle->setCheckoutUrl(
+            'https://automatic.fastbill.com/purchase/aa9122707e4baf2090e23babe7473a79/1'
+        );
         $expectedArticle->addFeature(
             new Feature('code', 2, 'value')
         );
-        $this->assertEquals(array($expectedArticle), $response->getArticles());
+
+        return $expectedArticle;
     }
 
     /**